# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VIP Tutors Task Management API",
        "description": "A comprehensive task management system with user authentication, task CRUD operations, reordering capabilities, and admin functionality.",
        "contact": {
            "name": "VIP Tutors Development Team",
            "email": "admin@viptutors.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get user's tasks",
                "description": "Retrieve a paginated list of tasks for the authenticated user with optional filtering and sorting",
                "operationId": "7786da811325606208249ebc8d393d9a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by task status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by priority (comma-separated for multiple)",
                        "schema": {
                            "type": "string",
                            "example": "high,medium"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search tasks by title or description",
                        "schema": {
                            "type": "string",
                            "example": "project"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter tasks created from this date",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter tasks created until this date",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "updated_at",
                                "title",
                                "priority",
                                "status",
                                "order"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (5-10)",
                        "schema": {
                            "type": "integer",
                            "maximum": 10,
                            "minimum": 5,
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "description": "Create a new task for the authenticated user",
                "operationId": "a737e990cef5769592bc41ee6a62f7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Complete project documentation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Write comprehensive documentation for the project"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ],
                                        "example": "medium"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "completed"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/tasks/reorder": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Reorder tasks",
                "description": "Reorder tasks for the authenticated user by providing an array of task IDs in the desired order",
                "operationId": "3c5d6dcecf0a0036597455c6500b6a53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tasks"
                                ],
                                "properties": {
                                    "tasks": {
                                        "description": "Array of task IDs in the desired order",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            3,
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tasks reordered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tasks reordered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or unauthorized task access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "description": "Task ID",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string",
                        "example": "Complete project documentation"
                    },
                    "description": {
                        "description": "Task description",
                        "type": "string",
                        "example": "Write comprehensive documentation for the project including API docs"
                    },
                    "status": {
                        "description": "Task status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ],
                        "example": "pending"
                    },
                    "priority": {
                        "description": "Task priority",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "medium"
                    },
                    "order": {
                        "description": "Task order for sorting",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user who owns this task",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Task creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Task last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}